+++ lib/ff_dpdk_kni.h
@@ -40,15 +40,15 @@ enum FilterReturn {
 void ff_kni_init(uint16_t nb_ports, const char *tcp_ports,
     const char *udp_ports);
 
-void ff_kni_alloc(uint8_t port_id, unsigned socket_id,
+void ff_kni_alloc(uint16_t port_id, unsigned socket_id,
     struct rte_mempool *mbuf_pool, unsigned ring_queue_size);
 
-void ff_kni_process(uint8_t port_id, uint16_t queue_id,
+void ff_kni_process(uint16_t port_id, uint16_t queue_id,
     struct rte_mbuf **pkts_burst, unsigned count);
 
 enum FilterReturn ff_kni_proto_filter(const void *data, uint16_t len);
 
-int ff_kni_enqueue(uint8_t port_id, struct rte_mbuf *pkt);
+int ff_kni_enqueue(uint16_t port_id, struct rte_mbuf *pkt);
 
 
 #endif /* ifndef _FSTACK_DPDK_KNI_H */
